+++ common/changelog.md
@@ -617,3 +617,5 @@
 - Mages - Removes DTNight set except for RDM as this has no functional use.
 - Moves the UI element base variables to the top of the gcdisplay file for idiots that are unable to just locate it 10 lines down.
 - Melee - Adds /sethp and related commands to allow control over regen HP values without reloading profiles
+- DRG - Remove redundant breath set and rename BreathBonus to HealingBreath_SteadyWing_SpiritLink
+- DRG - Remove pointless super jump set